requirements.txt
pandas
python-dotenv
pyodbc
SQLAlchemy
psycopg2
tqdm
cx-Oracle